requirements.txt
tensorflow
opencv-python
numpy
pillow
firebase-admin
